msm: thermal: Add support to online the hotplugged cores
Add support in KTM to online the hotplugged cores when the
thermal condition is cleared.

In newer targets, KTM has the responsibility of
onlining the cores hotplugged due to thermal
condition. This feature is enabled by defining a device
tree property. If this property is not defined, KTM
does the legacy functionality where it doesn't
bring the core back online.

Change-Id: I3c75466774b309846df3176df6029b886bd59896
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Conflicts:
	drivers/thermal/msm_thermal.c