soc: qcom: rq_stats: Always show idle time left
The def_timer_ms file indicates how much time has passed since we
wrote to the file to start the timer, but the time is only
updated when the work runs, instead of on demand when the file is
read. If we're poll(3)ing this file and another file and the
other file wakes our process up from sleep we won't know without
measuring time ourselves. Make this easier on userspace by always
showing the amount of time since we wrote the file so that we can
figure out how long we were asleep for.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Conflicts:
	drivers/soc/qcom/msm_rq_stats.c