MSM_Thermal: Modify the thermal steps.
The left bitwise shift in this equated to the following modifications to the temp threshold:
Hot = Temp threshold + 4 degress
Very Hot = temp threshold + 8 degrees
Hell = temp threshold + 16 degrees.

While the 4 and 8 are reasonable, +16 degrees for the last step is less reasonable.
Instead, make it step through the variables at a fixed value.

Ideally, the step value will be exposed to userspace.